requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coverage==5.5
Django==3.2
django-celery-beat==2.2.0
django-crispy-forms==1.11.2
django-timezone-field==4.1.2
djangorestframework==3.12.4
fontawesome-free==5.15.3
httpie==2.4.0
humanize==3.4.1
idna==2.10
kombu==4.6.11
prometheus-client==0.8.0
prompt-toolkit==3.0.18
Pygments==2.8.1
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
raven==6.10.0
requests==2.25.1
requests-toolbelt==0.9.1
sentry-sdk==1.0.0
six==1.15.0
sqlparse==0.4.1
tornado==6.1
Unipath==1.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
